+++ src/features/speedrun/season3/constants.ts
@@ -2,9 +2,9 @@ import { PlayerType } from "isaac-typescript-definitions";
 import g from "../../../globals";
 
 export const SEASON_3_CHARACTERS: readonly PlayerType[] = [
-  PlayerType.ISAAC, // 0
   PlayerType.JUDAS, // 3
   PlayerType.BLUE_BABY, // 4
+  PlayerType.SAMSON, // 6
   PlayerType.LAZARUS, // 8
   PlayerType.JUDAS_B, // 24
   PlayerType.EVE_B, // 26